factor out common class add travis file
