+++ build.sbt
@@ -17,17 +17,9 @@ lazy val jmhSettings = Seq(
   )
 )
 
-lazy val jsonBench = (project in file("json-bench"))
-  .enablePlugins(JmhPlugin)
-  .settings(jmhSettings)
+lazy val common = (project in file("common"))
   .settings(
-    name := "json-bench",
-    mainClass := Some("com.github.saint1991.samples.JsonBench"),
-    libraryDependencies ++= Seq(
-      "io.circe" %% "circe-core" % "0.9.2",
-      "io.circe" %% "circe-generic" % "0.9.2",
-      "io.circe" %% "circe-parser" % "0.9.2"
-    )
+    name := "common"
   )
 
 lazy val csvBench = (project in file("csv-bench"))
@@ -35,18 +27,48 @@ lazy val csvBench = (project in file("csv-bench"))
   .settings(jmhSettings)
   .settings(
     name := "csv-bench",
-    mainClass := Some("com.github.saint1991.samples.CsvBench")
+    mainClass := Some("com.github.saint1991.serialization.benchmark.FileGen")
   )
+  .dependsOn(common % "compile->compile")
+
+lazy val circeBench = (project in file("circe-bench"))
+  .enablePlugins(JmhPlugin)
+  .settings(jmhSettings)
+  .settings(
+    name := "json-bench",
+    mainClass := Some("com.github.saint1991.serialization.benchmark.FileGen"),
+    libraryDependencies ++= Seq(
+      "io.circe" %% "circe-core" % "0.9.3",
+      "io.circe" %% "circe-generic" % "0.9.3",
+      "io.circe" %% "circe-parser" % "0.9.3"
+    )
+  )
+  .dependsOn(common % "compile->compile")
 
 lazy val jsoniterScalaBench = (project in file("jsoniter-scala-bench"))
   .enablePlugins(JmhPlugin)
   .settings(jmhSettings)
   .settings(
     name := "jsoniter-scala-bench",
-    mainClass := Some("com.github.saint1991.samples.JsonBench"),
+    mainClass := Some("com.github.saint1991.serialization.benchmark.FileGen"),
     libraryDependencies ++= Seq(
       "com.github.plokhotnyuk.jsoniter-scala" %% "macros" % "0.26.0"
     )
+  )
+  .dependsOn(common % "compile->compile")
+
+lazy val avroBench = (project in file("avro-bench"))
+  .enablePlugins(JmhPlugin)
+  .settings(jmhSettings)
+  .settings(
+    name := "avro-bench",
+    mainClass := Some("com.github.saint1991.serialization.benchmark.FileGen"),
+    libraryDependencies += "org.apache.avro" % "avro" % "1.8.2"
+  )
+  .settings(
+    avroSpecificSourceDirectory in Compile := file("schema/avro"),
+    avroSpecificScalaSource in Compile := (sourceManaged in Compile).value,
+    sourceGenerators in Compile += (avroScalaGenerateSpecific in Compile).taskValue
   )
 
 lazy val protoBench = (project in file("proto-bench"))
@@ -54,11 +76,11 @@ lazy val protoBench = (project in file("proto-bench"))
   .settings(jmhSettings)
   .settings(
     name := "protobuf-bench",
-    mainClass := Some("com.github.saint1991.samples.ProtoBench"),
+    mainClass := Some("com.github.saint1991.serialization.benchmark.FileGen"),
   )
   .settings( // for ScalaPB
     PB.targets in Compile := Seq(scalapb.gen() -> (sourceManaged in Compile).value),
-    PB.protoSources in Compile := Seq(file("serialization/protocol-buffers"))
+    PB.protoSources in Compile := Seq(file("schema/protocol-buffers"))
   )
 
 lazy val thriftBench = (project in file("thrift-bench"))
@@ -66,37 +88,14 @@ lazy val thriftBench = (project in file("thrift-bench"))
   .settings(jmhSettings)
   .settings(
     name := "thrift-bench",
-    mainClass := Some("com.github.saint1991.samples.ThriftBench"),
+    mainClass := Some("com.github.saint1991.serialization.benchmark.FileGen"),
     libraryDependencies ++= Seq(
       "org.apache.thrift" % "libthrift" % "0.11.0",
-      "com.twitter" %% "scrooge-core" % "18.3.0" exclude("com.twitter", "libthrift")
+      "com.twitter" %% "scrooge-core" % "18.5.0" exclude("com.twitter", "libthrift")
     )
   )
   .settings(
-    scroogeThriftIncludeFolders in Compile := Seq(file("serialization/thrift")),
-    scroogeThriftSourceFolder in Compile := file("serialization/thrift"),
+    scroogeThriftIncludeFolders in Compile := Seq(file("schema/thrift")),
+    scroogeThriftSourceFolder in Compile := file("schema/thrift"),
     scroogeThriftOutputFolder in Compile := (sourceManaged in Compile).value
-  )
-
-lazy val avroBench = (project in file("avro-bench"))
-  .enablePlugins(JmhPlugin)
-  .settings(jmhSettings)
-  .settings(
-    name := "avro-bench",
-    mainClass := Some("com.github.saint1991.samples.AvroBench"),
-    libraryDependencies += "org.apache.avro" % "avro" % "1.8.2"
-  )
-  .settings(
-    avroSpecificSourceDirectory in Compile := file("serialization/avro"),
-    avroSpecificScalaSource in Compile := (sourceManaged in Compile).value,
-    sourceGenerators in Compile += (avroScalaGenerateSpecific in Compile).taskValue
-  )
-
-lazy val msgPackBench = (project in file("msgpack-bench"))
-  .enablePlugins(JmhPlugin)
-  .settings(jmhSettings)
-  .settings(
-    name := "msgpack-bench",
-    mainClass := Some("com.github.saint1991.samples.MsgPackBench"),
-    libraryDependencies += "org.msgpack" %% "msgpack-scala" % "0.8.13"
   )
