--- csv-bench/src/test/scala/com/github/saint1991/samples/CsvBench.scala
@@ -1,41 +1,0 @@
-package com.github.saint1991.samples
-
-import java.util.concurrent.TimeUnit
-
-import org.openjdk.jmh.annotations.{BenchmarkMode, Fork, Measurement, Mode, OutputTimeUnit, Scope, State, Warmup, Benchmark => JmhBenchmark}
-
-@State(Scope.Thread)
-@Warmup(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
-@Measurement(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
-@Fork(value = 1, jvmArgs = Array(
-  "-server",
-  "-Xms2g",
-  "-Xmx2g",
-  "-XX:NewSize=1g",
-  "-XX:MaxNewSize=1g",
-  "-XX:InitialCodeCacheSize=512m",
-  "-XX:ReservedCodeCacheSize=512m",
-  "-XX:+UseParallelGC",
-  "-XX:-UseBiasedLocking",
-  "-XX:+AlwaysPreTouch"
-))
-@OutputTimeUnit(TimeUnit.SECONDS)
-class CsvBench {
-  import Nobid._
-
-  final val N = 100000
-  val dataset = DataSet.createDataset(N)
-
-  val encodedDataset = encode()
-  decode()
-
-  @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
-  def encode(): Seq[String] = {
-    dataset.map(toCsv)
-  }
-
-  @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
-  def decode(): Seq[Nobid] = {
-    encodedDataset.map(fromCsv)
-  }
-}
