+++ jsoniter-scala-bench/src/test/scala/com/github/saint1991/serialization/benchmark/JsonIterScalaBench.scala
@@ -1,0 +1,42 @@
+package com.github.saint1991.serialization.benchmark
+
+import java.util.concurrent.TimeUnit
+
+import com.github.plokhotnyuk.jsoniter_scala.core._
+import com.github.plokhotnyuk.jsoniter_scala.macros._
+import org.openjdk.jmh.annotations.{BenchmarkMode, Fork, Measurement, Mode, OutputTimeUnit, Scope, State, Warmup, Benchmark => JmhBenchmark}
+
+import com.github.saint1991.serialization.benchmark.dataset._
+
+@State(Scope.Thread)
+@Warmup(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
+@Measurement(iterations = 10, time = 1, timeUnit = TimeUnit.SECONDS)
+@Fork(value = 1, jvmArgs = Array(
+  "-server",
+  "-Xms2g",
+  "-Xmx2g",
+  "-XX:NewSize=1g",
+  "-XX:MaxNewSize=1g",
+  "-XX:InitialCodeCacheSize=512m",
+  "-XX:ReservedCodeCacheSize=512m",
+  "-XX:+UseParallelGC",
+  "-XX:-UseBiasedLocking",
+  "-XX:+AlwaysPreTouch"
+))
+@OutputTimeUnit(TimeUnit.SECONDS)
+class JsonIterScalaBench {
+  final val N = 100000
+  val dataset: Seq[Nobid] = DataSet.createDataset(N)
+
+  implicit val codec: JsonValueCodec[Nobid] = JsonCodecMaker.make[Nobid](CodecMakerConfig())
+
+  val encodedDataset: Seq[Array[Byte]] = encode()
+  decode()
+
+  @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
+  def encode(): Seq[Array[Byte]] = dataset.map(x => writeToArray(x))
+
+  @JmhBenchmark @BenchmarkMode(Array(Mode.AverageTime))
+  def decode(): Seq[Nobid] = encodedDataset.map(str => readFromArray(str))
+}
+
