+++ project/plugins.sbt
@@ -1,6 +1,6 @@
-addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.3.3")
+addSbtPlugin("pl.project13.scala" % "sbt-jmh" % "0.3.4")
 addSbtPlugin("com.thesamet" % "sbt-protoc" % "0.99.18")
-addSbtPlugin("com.julianpeeters" % "sbt-avrohugger" % "2.0.0-RC4")
-addSbtPlugin("com.twitter" % "scrooge-sbt-plugin" % "18.3.0")
+addSbtPlugin("com.julianpeeters" % "sbt-avrohugger" % "2.0.0-RC9")
+addSbtPlugin("com.twitter" % "scrooge-sbt-plugin" % "18.5.0")
 
 libraryDependencies += "com.thesamet.scalapb" %% "compilerplugin" % "0.7.0"
