+++ schema/avro/nobid.avsc
@@ -1,0 +1,41 @@
+{
+    "namespace": "com.github.saint1991.serialization.benchmark",
+    "name":"Nobid",
+    "type":"record",
+    "fields":[
+        {"name":"adnwId","type":"int"},
+        {"name":"appName","type":"string"},
+        {"name":"auctionId","type":"string"},
+        {"name":"host","type":"string"},
+        {"name":"loggedAt","type":"string"},
+        {"name":"mId","type":"int"},
+        {"name":"nbr","type":"int"},
+        {"name":"page","type":["null", "string"], "default": null},
+        {"name":"resTime","type":"int"},
+        {"name":"spot","type": {
+            "name": "spotRecord",
+            "type": "record",
+            "fields": [
+                {"name": "id", "type": "int"},
+                {"name": "type", "type": {
+                    "name": "spotType",
+                    "type": "enum",
+                    "symbols": ["A", "S"]
+                }}
+            ]
+        }},
+        {"name": "history", "type": {
+            "name": "historyItems",
+            "type": "array",
+            "items": {
+                "name": "historyItem",
+                "type": "string"
+            }
+        }},
+        {"name": "tags", "type": {
+            "name": "tag",
+            "type": "map",
+            "values": "string"
+        }}
+    ]
+}
