+++ schema/protocol-buffers/nobid.proto
@@ -1,0 +1,40 @@
+syntax = "proto3";
+package benchmark;
+option java_package = "com.github.saint1991.serialization.benchmark";
+
+// SPEED (default) | CODE_SIZE | LITE_RUNTIME
+option optimize_for = SPEED;
+
+// possible to import definition from other files
+import "spot.proto";
+
+// message should be CamelCase
+message Nobid {
+
+  // fields should be snake_case
+  // each field should have an unique index
+  // required directive indicate a field must exist
+  int32  adnw_id = 1;
+  string app_name = 2;
+  string auction_id = 3;
+  string host = 4;
+  string logged_at = 5;
+  int32  m_id = 6;
+  int32  nbr = 7;
+  string page = 8;
+  int32  res_time = 9;
+  benchmark.Spot spot = 10;
+
+
+  repeated string history = 11;
+  map<string, string> tags = 12;
+
+  // it's possible to reserve some indices for future use
+  reserved 13 to 15;
+}
+
+message Response {
+  int32 status_code = 1;
+  string message = 2;
+}
+
