--- serialization/thrift/spot.thrift
@@ -1,12 +1,0 @@
-
-namespace java com.github.saint1991.samples
-
-enum SpotType {
-    A,
-    S
-}
-
-struct Spot {
-    1:required i32      id
-    2:required SpotType type
-}
